+++ CHANGELOG.md
@@ -12,6 +12,9 @@ DEPRECATIONS/CHANGES:
    `disallowed_policies` in role definitions in the token auth backend, input
    can now be a comma-separated string or an array of strings. Reading a role
    will now return arrays for these parameters.
+ * Transit key exporting: You can now mark a key in the `transit` backend as
+   `exportable` at any time, rather than just at creation time; however, once
+   this value is set, it still cannot be unset.
  * PKI Secret Backend Roles parameter types: For `allowed_domains` and
    `key_usage` in role definitions in the PKI secret backend, input
    can now be a comma-separated string or an array of strings. Reading a role